+++ CHANGELOG.md
@@ -17,12 +17,12 @@
     - Styling improvements of all sheets
     - Added Character Details and Contacts sections that were missing
     - Changed limited sheets to only show minimal information (Image, Name, Unrestricted Bio/Description)
+    - Improved chat message rolls (Reorganized, color-coded, etc..)
 - Sheet templates overhaul: All sheets are now using modular templates, preventing unnecessary repetitions
 - Organized character template data: This is a breaking change for characters and creatures as the attributes are now nested under "primaryAttributes" and "secondaryAttributes" instead of at the root level of the character data and the description box content is now under "bio.biography" alongside all of the other character details
 - Removed Aspects selection from creature rolls as these do not provide roll modifiers
 - Removed unused field from vehicle weapons
 - Added missing damage type field to vehicle weapon rolls
-- Improved chat message rolls (Reorganized, color-coded, etc..)
 - Added dynamically generated flavor text in several places